Tries to mitigate the speed regression introduced by #18. The ports are now initialized before the model so that changes do not trigger unnecessary functions anymore.
Also removes usages of `portProp`.

Benchmarking gave the following results:
- serialadder:
  - before-ports: 2.9 s
  - after-ports: 4 s
  - with speedup: 3.5 s
- horner:
  - before-ports: 15.5 s
  - after-ports: 70 s
  - with speedup: 25 s

So, unfortunately, even after trying to remove unnecessary function calls, there is still a speed loss compared to before using the jointJS port API. I fear this might be inherent due to some logic overhead to handle ports in jointJS.