@tilk 2020-06-23
The slowdown now feels more acceptable. Maybe there will be a way later to reclaim this little bit of lost performance.